requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
PyPDF2
python-dotenv
groq
sentence-transformers